Add `info` type for Alert component
There is a slight confusion about default/info currently (code uses
default but text mentions info).

In addition to success/warning/error types, it makes sense to have two
more where one is shallow and the other emphasized.

`default` is currently blending in with the background as it's light
gray. The one this PR adds (`info`) would be emphasized. I took the
already existing status color for the `info` type.