Submit timestamp value only on enter or blur
Previously, when editing the timestamp value, the input field value was
replaced as soon as a loosely parsable date was found. This made editing
the timestamp cumbersome.

With this PR, the date will only be checked for validity and
subsequently sent to its parent only if Enter is pressed or the input loses
focus.