only render children when the dialog is active
Due to being hidden with css only, and still added to the DOM, any child
component is instantiated when the containing page is rendered.
If you have  a Dialog child with any heavy computation, or any at all
for that matter, you would be doing it even though the dialog may never
be opened.
This commit changes that by only rendering the children when the dialog is
active.

fixes #357