fix: adds engines field to package.json
see: https://nodejs.org/en/about/releases/

- Node 8 LTS (released 2017-05-30) is in Maintenance status, not in Active status.
- Node 10 LTS (released 2018-02-24) is in Active status.
- Node 12 LTS (released 2019-04-23) is in Current status.

I built this project with 12.10.0 initially, but it didn't work with many pages of error messages it took me a second to read through (compared to the clean error you get when there's an `engines` field in the `package.json`).  I tried with 8.16.1 and it worked.  I tried with 10.16.3 and it worked.  Therefore I added an `engines` field to the `package.json` with ">=8 <12".